+++ DriveScrewMotionSystem.h
@@ -12,28 +12,28 @@
 class DriveScrewMotionSystem: public MotionSystem{
   private:
   
-  float travelLength;
-  float numberStepsPerTurn;
-  float microstepping;
-  float pitch;
-  float gearRatio;
+  double travelLength;
+  double numberStepsPerTurn;
+  double microstepping;
+  double pitch;
+  double gearRatio;
   
   
   public:
-  DriveScrewMotionSystem(float travelLength, unsigned int numberStepsPerTurn, unsigned int microstepping, float diameter, float gearRatio);
+  DriveScrewMotionSystem(double travelLength, unsigned int numberStepsPerTurn, unsigned int microstepping, double diameter, double gearRatio);
 
   
    /**
    *@param pos the position in milimeters that needs to be converted to a stepCount;
    *@return the number of micro steps needed for the system to move by 1mm.
    */ 
-  unsigned long convertPosToStepCount(float pos);
+  unsigned long convertPosToStepCount(double pos);
 
   /**
    * @param stepCount the number of steps that needs to be converted to a pos in milimeters
    */
-  float convertStepCountToPos(unsigned long stepCount);
-
+  double convertStepCountToPos(unsigned long stepCount);
+  
   /**
    * @return the maximum step position given the motion systems travel length
    */
